#69 Feat: 유저프로필 상단 컴포넌트화
