+++ iTER/src/App.tsx
@@ -14,6 +14,7 @@ import Profile from './pages/user/Profile';
 import Onboading from './pages/Onboading';
 import Interest from './pages/mypage/Interest';
 import Like from './pages/mypage/Like';
+import Index from './pages/mypage/Index';
 
 function App() {
   return (
@@ -31,6 +32,7 @@ function App() {
       <Route path="/search/review/:id" element={<ReviewDetail />} />
       <Route path="/user/profile/:id" element={<Profile />} />
       <Route path="/onboading" element={<Onboading />} />
+      <Route path="/mypage" element={<Index />} />
       <Route path="/mypage/interest" element={<Interest />} />
       <Route path="/mypage/like" element={<Like />} />
     </Routes>
